+++ src/clj_kondo/impl/linters.clj
@@ -219,6 +219,20 @@
          :single-logical-operand
          (format "Single arg use of %s always returns the arg itself" call-name))))))
 
+(defn lint-redundant-nested-call
+  "Lints calls of variadic functions/macros when nested."
+  [call]
+  (let [[[call-ns call-name] parent] (:callstack call)]
+    (when (and (utils/one-of call-ns [clojure.core cljs.core])
+               (utils/one-of call-name [* *' + +' and comp concat every-pred
+                                        lazy-cat max merge min or some-fn str])
+               (= [call-ns call-name] parent))
+      (node->line
+       (:filename call)
+       (:expr call)
+       :redundant-nested-call
+       (format "Redundant nested call: %s" call-name)))))
+
 #_(require 'clojure.pprint)
 
 (defn lint-var-usage
@@ -297,9 +311,9 @@
                                   (when called-fn
                                     (or different-file?
                                         (not row-called-fn)
-                                        (or (> row-call row-called-fn)
-                                            (and (= row-call row-called-fn)
-                                                 (> (:col call) (:col called-fn)))))))
+                                        (> row-call row-called-fn)
+                                        (and (= row-call row-called-fn)
+                                             (> (:col call) (:col called-fn))))))
                   _ (when-let [t (:type called-fn)]
                       (when (and call? (utils/one-of t [:string]))
                         (findings/reg-finding!
@@ -401,7 +415,11 @@
                   single-logical-operand-error
                   (and call?
                        (not (utils/linter-disabled? call :single-logical-operand))
-                       (lint-single-logical-operand call))]]
+                       (lint-single-logical-operand call))
+                  redundant-nesting-error
+                  (and call?
+                       (not (utils/linter-disabled? call :redundant-nested-call))
+                       (lint-redundant-nested-call call))]]
       (when (and (not call?)
                  (identical? :fn (:type called-fn)))
         (when (:condition call)
@@ -422,6 +440,9 @@
       (when single-logical-operand-error
         (findings/reg-finding! ctx
                                single-logical-operand-error))
+      (when redundant-nesting-error
+        (findings/reg-finding! ctx
+                               redundant-nesting-error))
       (when (and (:private called-fn)
                  (not= caller-ns-sym
                        fn-ns)